Add support for sending mail via SMTP
This is useful for users who use mailcatcher to test and visualize email
locally.

This change introduces 2 additional parameters in
`Sendgrid::Mailer.initialize`

1. `env`
  This is to replace `Rails.env.production?` hence decouples
from having `Rails` as dependencies. This is not a breaking change as it
will be default to `:production`.

2. `delivery_method`

There are 2 delivery methods for now, `:api` and `:smtp`. By using
`:api`, the email will be sent via SendGrid API. By
using `:smtp`, the email will be sent via SMTP using the `Mail` gem,
hence the additional runtime dependency for this gem.